Change ChatService to use KittehIRC
