+++ src/main/java/com/faforever/client/chat/ChatChannel.java
@@ -1,19 +1,16 @@
 package com.faforever.client.chat;
 
 import com.faforever.client.fx.JavaFxUtil;
-import javafx.beans.property.ReadOnlySetWrapper;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.MapChangeListener;
 import javafx.collections.ObservableMap;
-import javafx.collections.ObservableSet;
 import lombok.Value;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 @Value
 public class ChatChannel {
@@ -21,12 +18,10 @@ public class ChatChannel {
   ObservableMap<String, ChatChannelUser> users;
   StringProperty topic;
   String name;
-  ObservableSet<String> moderators;
 
   public ChatChannel(String name) {
     this.name = name;
     users = FXCollections.synchronizedObservableMap(FXCollections.observableHashMap());
-    moderators = FXCollections.observableSet();
     topic = new SimpleStringProperty();
   }
 
@@ -51,9 +46,6 @@ public class ChatChannel {
   }
 
   public void addUser(ChatChannelUser user) {
-    if (moderators.contains(user.getUsername())) {
-      user.setModerator(true);
-    }
     users.put(user.getUsername(), user);
   }
 
@@ -67,24 +59,6 @@ public class ChatChannel {
 
   public void removeUserListener(MapChangeListener<String, ChatChannelUser> listener) {
     users.removeListener(listener);
-  }
-
-  public void addModerator(String username) {
-    Optional.ofNullable(users.get(username)).ifPresent(user -> user.setModerator(true));
-    synchronized (moderators) {
-      moderators.add(username);
-    }
-  }
-
-  public void removeModerator(String username) {
-    Optional.ofNullable(users.get(username)).ifPresent(user -> user.setModerator(false));
-    synchronized (moderators) {
-      moderators.remove(username);
-    }
-  }
-
-  public ReadOnlySetWrapper<String> getModerators() {
-    return new ReadOnlySetWrapper<>(moderators);
   }
 
   /**
