--- src/main/java/com/faforever/client/chat/PircBotXFactory.java
@@ -1,9 +1,0 @@
-package com.faforever.client.chat;
-
-import org.pircbotx.Configuration;
-import org.pircbotx.PircBotX;
-
-public interface PircBotXFactory {
-
-  PircBotX createPircBotX(Configuration configuration);
-}
